requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
supabase==2.10.0
sentence-transformers==3.3.1
langchain-community==0.3.13
langchain-text-splitters==0.3.4
pypdf==5.1.0
rank-bm25==0.2.2
groq==0.13.1
ragas==0.2.8
python-dotenv==1.0.1

